+++ memoria/include/memoria_config.h
@@ -1,8 +1,9 @@
 #ifndef MEMORIA_CONFIG
 #define MEMORIA_CONFIG
 
-    #include "networking.h"
-    #include "serializacion.h"
+    #include <commons/config.h>
+    #include <netdb.h>
+    #include <stdlib.h>
 
     typedef struct
     {
@@ -10,8 +11,8 @@
         char* ip_memoria;
         char* puerto_escucha; // Puerto por el cual se escuchará la conexión de módulo
         uint32_t tamanio_memoria; // Tamaño expresado en bytes
-        uint8_t tamanio_pagina; // Tamaño de las páginas en bytes
-        uint8_t paginas_por_tabla; // Cantidad de entradas de cada tabla de páginas
+        uint32_t tamanio_pagina; // Tamaño de las páginas en bytes
+        uint32_t paginas_por_tabla; // Cantidad de entradas de cada tabla de páginas
         int retardo_memoria; // Tiempo en milisegundos que se deberá esperar para dar una respuesta al CPU
         char* algoritmo_reemplazo; // Algoritmo de reemplazo de páginas (CLOCK/CLOCK-M)
         int marcos_por_proceso; // Cantidad de marcos permitidos por proceso en asignación fija
